+++ src/metrics/detection/heatmap/heatmap_average_precision.py
@@ -35,12 +35,12 @@ class HeatmapAveragePrecision:
 
     def calculate(
         self,
-        pred_lines_batch: List[ArrayNx4[np.float]],
-        gt_lines_batch: List[ArrayNx4[np.float]],
-        scores_batch: List[ArrayN[np.float]],
-        heights_batch: ArrayN[np.int],
-        widths_batch: ArrayN[np.int],
-        thresholds: ArrayN[np.int],
+        pred_lines_batch: List[ArrayNx4[float]],
+        gt_lines_batch: List[ArrayNx4[float]],
+        scores_batch: List[ArrayN[float]],
+        heights_batch: ArrayN[int],
+        widths_batch: ArrayN[int],
+        thresholds: ArrayN[int],
     ):
         precision, recall = self.precision_recall_calculator.calculate(
             pred_lines_batch,
@@ -56,12 +56,12 @@ class HeatmapAveragePrecision:
 
 
 def heatmap_average_precision(
-    pred_lines_batch: List[ArrayNx4[np.float]],
-    gt_lines_batch: List[ArrayNx4[np.float]],
-    line_scores_batch: List[ArrayNx4[np.float]],
-    heights_batch: ArrayN[np.int],
-    widths_batch: ArrayN[np.int],
-    thresholds: ArrayN[np.int],
+    pred_lines_batch: List[ArrayNx4[float]],
+    gt_lines_batch: List[ArrayNx4[float]],
+    line_scores_batch: List[ArrayNx4[float]],
+    heights_batch: ArrayN[int],
+    widths_batch: ArrayN[int],
+    thresholds: ArrayN[int],
 ):
     """
     Calculates Heatmap Average Precision (AP^H)
