+++ src/metrics/detection/vectorized/structural_average_precision.py
@@ -12,8 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import numpy as np
-
 from typing import List
 
 from src.metrics.detection.vectorized.average_precision import AveragePrecision
@@ -27,9 +25,9 @@ __all__ = ["structural_average_precision"]
 
 
 def structural_average_precision(
-    pred_lines_batch: List[ArrayNx4[np.float]],
-    gt_lines_batch: List[ArrayNx4[np.float]],
-    line_scores_batch: List[ArrayN[np.float]],
+    pred_lines_batch: List[ArrayNx4[float]],
+    gt_lines_batch: List[ArrayNx4[float]],
+    line_scores_batch: List[ArrayN[float]],
     distance_threshold: float = 5,
 ) -> float:
     """
